added and modified scripts

Former-commit-id: cdb7da34467c07955e934d390e878dac0f0d5fb3

+++ A_Basic/06_Modules/04_argparse/argparsedemo.py
@@ -20,6 +20,6 @@ if __name__ == '__main__':
                         required=True)
     parser.add_argument('-c',
                         help='Third value',
-                        type=float)
+                        type=int)
     args = parser.parse_args()
     print(main(args.a, args.b, args.c))
